Optimized bucketed random variable.
Bucketed random variable kept a long "array" which could cause state
explosion especially after convolutions.  An array with min value of 0
convoled by another array with min value of 10000 could result in an
array with 10k+ entries.

Changed the array of rvar_type_t's to array of bucket_t which besides
the probability also keeps the value of the bucket.  This way we can
avoid zero buckets.